+++ pyworkflow/em/packages/pkpd/protocol_pkpd_two_compartments_clint.py
@@ -26,30 +26,24 @@
 
 import pyworkflow.protocol.params as params
 from protocol_pkpd_ode_base import ProtPKPDODEBase
-from pk_models import PK_Twocompartments
-import pyworkflow.em.biopharmaceutics
+from pk_models import PK_TwocompartmentsClint
 
-# TESTED in test_workflow_gabrielsson_pk07.py
-# TESTED in test_workflow_gabrielsson_pk08.py
-# TESTED in test_workflow_gabrielsson_pk09.py
-# TESTED in test_workflow_gabrielsson_pk10.py
-# TESTED in test_workflow_gabrielsson_pk11.py
-# TESTED in test_workflow_gabrielsson_pk12.py
-# TESTED in test_workflow_gabrielsson_pk13.py
-# TESTED in test_workflow_gabrielsson_pk14.py
-# TESTED in test_workflow_gabrielsson_pk16.py
-# TESTED in test_workflow_gabrielsson_pk39.py
+# TESTED in test_workflow_gabrielsson_pk18.py
 
-class ProtPKPDTwoCompartments(ProtPKPDODEBase):
+class ProtPKPDTwoCompartmentsClint(ProtPKPDODEBase):
     """ Fit a two-compartmentx model to a set of measurements (any arbitrary dosing regimen is allowed)\n
         The central compartment is referred to as C, while the peripheral compartment as Cp.
-        The differential equation is V dC/dt = -(Cl+Clp) * C + Clp * Cp + dD/dt, Vp dCp/dt = Clp * C - Clp * Cp\n
+        The differential equation is V dC/dt = -(Clint+Clp) * C + Clp * Cp + dD/dt, Vp dCp/dt = Clp * C - Clp * Cp\n
+        and Clint=Vmax/(Km+C)\n
         where C is the concentration of the central compartment, Cl the clearance, V and Vp the distribution volume of the central and peripheral compartment,
-        Clp is the distribution rate between the central and the peripheral compartments, and D the input dosing regime.
+        Clp is the distribution rate between the central and the peripheral compartments, \n
+        Vmax is the maximum processing capability of the metabolic pathway degrading the drug, \n
+        Km is the Michaelis-Menten constant,and D the input dosing regime. \n
+        Note that the intrinsic clearance occurs at the central volume.\n
         Confidence intervals calculated by this fitting may be pessimistic because it assumes that all model parameters
         are independent, which are not. Use Bootstrap estimates instead.\n
         Protocol created by http://www.kinestatpharma.com\n"""
-    _label = 'pk two-compartments'
+    _label = 'pk two-compartments intrinsic'
 
     def __init__(self,**kwargs):
         ProtPKPDODEBase.__init__(self,**kwargs)
@@ -57,10 +51,10 @@ class ProtPKPDTwoCompartments(ProtPKPDODEBase):
     #--------------------------- DEFINE param functions --------------------------------------------
     def _defineParams(self, form):
         self._defineParams1(form, True, "t", "Cp")
-        form.addParam('bounds', params.StringParam, label="Parameter bounds ([tlag], sourceParameters, Cl, V, Clp, Vp)", default="",
-                      help="Bounds for time delay, central clearance and volume and peripheral clearance and volume. "\
+        form.addParam('bounds', params.StringParam, label="Parameter bounds ([tlag], sourceParameters, Vmax, Km, V, Clp, Vp)", default="",
+                      help="Bounds for time delay, maximum processivity, Michaelis constant, volume and peripheral clearance and volume. "\
                       'Make sure that the bounds are expressed in the expected units (estimated from the sample itself).'\
                       'Be careful that Cl bounds must be given here. If you have an estimate of the elimination rate, this is Ke=Cl/V. Consequently, Cl=Ke*V ')
 
     def createModel(self):
-        return PK_Twocompartments()
+        return PK_TwocompartmentsClint()